Correction of the polarization bug found by Gaspard
The test on lisotropic was not done correctly.
itheta was set to 1 in all cases, resulting in a wrong selection of
the polarisability (and phase function value too)
The RT was not affected in single scattering as polarization is calculated
independently. The bug was however likely to affect RT in multiple scattering.

There is still a sign difference between MC and RT on the Q parameter