Improved resampling of stellar spectra
The average spectrum in the bin is selected only is the number of data
points is sufficient and if the average wavelength is not too far from
the center of the bin (this can happen because the sampling of the
spectra is not uniform). Otherwise, MCFOST interpolates.

The long wavelength regime is now a power-law extrapolation to avoid a shift
due to line-blanketing