Bug correction: PAH in ray-tracing mode with multiple zones
The wavelength index was not passed correctly to compute the wavelength bins in mueller_PAH. The bug was appearing in ray-tracing mode with multiple zones because scattering_method was set to 2 and p_lambda was pointing to 1. The resulting opacities and scattering fonction were wrong.

Also added a warning message when reaching the iteration limit for PAHs.

Minor correction of a bug in mem.f90 (array allocated twice in some configurations).

Truncated long lines in stars.f90.