Using the same pola interpolation for the single scattered component
Note: this is not as critical as for the mutiple scattered component.
There is a much higher resolution (nang_ray_tracing_star >> nang_ray_tracing).
The previous interpolation method altough not "correct" was giving correct
results too (thanks to the high angular res).
The new method is slower: the ray-tracing takes about 30% time more.
I keep it, as it is potentially better and only affects the cpu time for
polarised images.